Separate out grid interaction functions
